Moving more static parts to _document.js
